move tests to separated file for more clarity
